backend/pyproject.toml
[build-system]
requires = ["setuptools>=18.5", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "large-document-rag"
version = "0.1.0"
authors = [
  { name = "Adam Cobb", email = "adamgcobb@gmail.com" },
]
description = "RAG system for large document processing"
readme = "./README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    # Core dependencies
    "langchain==0.3.23",
    "langchain-community==0.3.21",
    "langchain_anthropic==0.3.10",
    "pydantic>=2.7.4,<3.0.0",
    "pydantic-settings>=2.2.0,<3.0.0",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "scikit-learn>=1.4.1",
    "scipy>=1.15.2",
    "tqdm>=4.67.1",

    # Vector storage
    "chromadb>=0.4.22",
    "hnswlib>=0.8.0",

    # Document processing
    "unstructured>=0.17.2",
    "pypdf>=5.4.0",
    "pytesseract>=0.3.13",
    "pdf2image>=1.16.3",
    "python-docx>=1.1.2",

    # Financial NLP/NER and ML
    "spacy>=3.8.5",

    # Custom Embeddings and Retrieval
    "cohere>=4.37",
    "scikit-learn-extra>=0.3.0",
    "scikit-multilearn>=0.2.0",

    # LLM integration
    "openai>=1.12.0",
    "anthropic>=0.49.0",
    "langchain-anthropic>=0.1.2",
    "transformers>=4.50.3",
    "transformers-interpret>=0.10.0",
    "sentence-transformers>=4.0.2",
    "torch>=2.6.0",

    # API and serving
    "fastapi==0.115.9",
    "uvicorn>=0.34.0",
    "fastapi-cache2>=0.2.1",
    "fastapi-utils>=0.2.1",
    "python-multipart>=0.0.20",

    # Multiprocessing and parallelization
    "joblib==1.4.2",
    "ray>=2.44.1",

    # Utilities
    "python-dotenv>=1.1.0",
    "pandas>=2.2.3",
    "pytz>=2025.2",
    "rich>=13.7.1",
    "tabulate>=0.9.0",
    
    # NLP and ML
    "spacy>=3.8.0",
    "transformers>=4.30.0",
    "huggingface_hub[hf_xet]>=0.30.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
large-document-rag-backend = "src.main:main"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.dynamic]
version = {attr = "src.__VERSION__"}
readme = {file = ["README.md"]}

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
